+++ lib/vagrant/errors.rb
@@ -158,6 +158,11 @@ module Vagrant
       error_key(:environment_locked)
     end
 
+    class HomeDirectoryMigrationFailed < VagrantError
+      status_code(53)
+      error_key(:home_dir_migration_failed)
+    end
+
     class ForwardPortAutolistEmpty < VagrantError
       status_code(27)
       error_key(:auto_empty, "vagrant.actions.vm.forward_ports")