+++ test/test_helper.rb
@@ -13,6 +13,9 @@ begin
 rescue LoadError
 end
 
+# Set the home directory to some temporary directory
+ENV["HOME"] = Vagrant.source_root.join("test", "tmp", "home").to_s
+
 # Add the I18n locale for tests
 I18n.load_path << File.expand_path("../locales/en.yml", __FILE__)